refactor: move geocoding logic to a separate utility file and simplify NavigationInput component
